requirements.txt
fastapi
uvicorn
langchain
langchain-openai
faiss-cpu
pydantic
python-dotenv
streamlit
tiktoken
python-dotenv
